requirements.txt
Flask==3.0.0
requests==2.31.0
Werkzeug==3.0.1
python-dotenv==1.0.0
Flask-Limiter==3.5.0
urllib3>=2.0.0
# Upgraded supabase to resolve httpx conflict
supabase>=2.3.0
# Let pip determine the compatible version
postgrest
# Relaxed constraint
httpx>=0.27.0
protobuf>=5.28.0
h11>=0.16.0
# JWT Authentication
PyJWT>=2.8.0
# Google OAuth Verification
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
flask
flask-limiter
python-dotenv
requests
supabase
gunicorn
# Markdown to HTML conversion for email templates
markdown